split lib into sub mod
